Update README.md
This Java program is designed to simulate a voting process based on data stored in an Excel file. It prompts the user to enter their name, age, and ID number. Then, it compares this data with the information stored in the Excel file. If the provided data matches with any record in the Excel file, the user is given the choice to cast their vote.